fix(perturbation): upgrade reference orbit precision to double
- Modified ComplexFloat struct in perturbation.h to use double instead of float.
- Updated perturbation_compute in perturbation.c to store orbit data as doubles.
- This is required for the new Dekker split arithmetic to accurately encode double precision coords into high/low floats for the GPU shader.

+++ include/perturbation.h
@@ -12,8 +12,8 @@
 
 // single-precision complex number format for gpu texture compatibility
 typedef struct {
-    float re;
-    float im;
+    double re;
+    double im;
 } ComplexFloat;
 
 // represents the stored reference orbit path for the center coordinate
